requirements.txt
boto3==1.9.211
botocore==1.12.211

mysql-connector-python==8.0.17 
pandas==0.25.3
requests==2.22.0
Django==2.2.4
djangorestframework==3.11.0
django-environ==0.4.5
markdown==3.2.1
django-filter==2.2.0

vaderSentiment==3.3.2
